% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_performance_results}
\alias{example_performance_results}
\title{Example performance analysis results}
\format{
An object of class \code{data.frame} with 12 rows and 7 columns.
}
\usage{
example_performance_results
}
\description{
Example performance analysis results
}
\keyword{datasets}
