% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mt_block_data}
\alias{mt_block_data}
\title{Example block-level population data from Montana for split precinct analysis}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3880 rows and 2 columns.
}
\usage{
mt_block_data
}
\description{
Example block-level population data from Montana for split precinct analysis
}
\keyword{datasets}
