% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_intersect.R
\name{percent_intersect}
\alias{percent_intersect}
\title{Calculate percent land area intersections}
\usage{
percent_intersect(sfdf, shp)
}
\arguments{
\item{sfdf}{sf dataframe with one or more geometries.}

\item{shp}{sf dataframe with a single shape boundary.}
}
\value{
sfdf dataframe with pct_intersect column
}
\description{
Calculates the percent area 
intersection between the geometries in a sf data.frame and a 
single boundary shape.
}
\author{
Rachel Carroll <rachelcarroll4@gmail.com>
}
