% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planShp}
\alias{planShp}
\title{Example district plan shape for split precinct analysis}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 18 columns.
}
\usage{
planShp
}
\description{
Example district plan shape for split precinct analysis
}
\keyword{datasets}
