% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vtd}
\alias{vtd}
\title{Example vtd-level sf dataframe with election results for split precinct analysis}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 19 rows and 12 columns.
}
\usage{
vtd
}
\description{
Example vtd-level sf dataframe with election results for split precinct analysis
}
\keyword{datasets}
