% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wa_block_data}
\alias{wa_block_data}
\title{Example block-level population data from Washington for BISG}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 821 rows and 8 columns.
}
\usage{
wa_block_data
}
\description{
Example block-level population data from Washington for BISG
}
\keyword{datasets}
