% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wa_geocoded}
\alias{wa_geocoded}
\title{Example geocoded voter file from Washington for BISG}
\format{
An object of class \code{data.frame} with 1000 rows and 6 columns.
}
\usage{
wa_geocoded
}
\description{
Example geocoded voter file from Washington for BISG
}
\keyword{datasets}
