% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{washington}
\alias{washington}
\title{Example election data with BISG demographics from Washington 2020 General Presidential Election}
\format{
An object of class \code{data.frame} with 637 rows and 27 columns.
}
\usage{
washington
}
\description{
Example election data with BISG demographics from Washington 2020 General Presidential Election
}
\keyword{datasets}
